Use string literal in place of regex literal
In a multibyte-safe configuration, GNU Awk fails to parse a literal
regex pattern with that uses octal escapes for bytes, failing with
"Invalid collation character: /^([-][-][ -~]*)+/". To prevent this from
happening, the regex literal has been replaced with a string literal
which presumably defers parsing until the line is actually executed
which should not happen if GNU Awk is running a multibyte-safe
configuration. This change seems to have no significant impact on
performance for most implementations with the sole exception of BusyBox
AWK which takes about 30% longer to process the French corpus and 8%
longer to process the Chinese corpus.

This commit resolves https://github.com/ericpruitt/wcwidth.awk/issues/2.