Update Unicode width tables
This updates the Unicode width tables to current values based on GNU
libc version 2.36-9 on Debian 12. The width of the combining characters
U+3099 and U+309A in the updated width table have a width of 0 which
necessitated amending the CJK optimization literal regex for
multibyte-safe AWK implementations and the corresponding regex
"WCWIDTH_WIDE_CJK_RUNES_REGEX" for implementations that are not
multibyte-safe.